backend/pyproject.toml
[project]
name = "fixit-chat-backend"
version = "0.1.0"
description = "FixIT Chat - техподдержка для fixitmail.ru"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "redis[hiredis]>=5.0,<6",
    "bcrypt>=4.0,<6",
    "python-jose[cryptography]>=3.3,<4",
    "python-multipart>=0.0.18",
    "python-magic>=0.4,<1",
    "cryptography>=44,<45",
    "structlog>=24,<26",
    "httpx>=0.27,<1",
    "apscheduler>=3.10,<4",
    "prometheus-client>=0.21,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
    "factory-boy>=3.3,<4",
    "ruff>=0.8,<1",
    "coverage>=7,<8",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
